# pd-client: implement GetTSAsync use streaming