--- a/cmd/pd-tso-bench/main.go
+++ b/cmd/pd-tso-bench/main.go
@@ -17,6 +17,8 @@ import (
 	"flag"
 	"fmt"
 	"log"
+	"net/http"
+	_ "net/http/pprof"
 	"os"
 	"os/signal"
 	"sync"
@@ -72,6 +74,10 @@ func main() {
 	go func() {
 		<-sc
 		cancel()
+	}()
+
+	go func() {
+		http.ListenAndServe(":6666", nil)
 	}()
 
 	wg.Wait()