--- a/pd-client/client.go
+++ b/pd-client/client.go
@@ -14,6 +14,7 @@
 package pd
 
 import (
+	"fmt"
 	"net"
 	"net/url"
 	"strings"
@@ -348,6 +349,8 @@ func (c *client) tsLoop() {
 	}
 }
 
+var global uint64
+
 func (c *client) processTSORequests(stream pdpb.PD_TsoClient, requests []*tsoRequest) error {
 	start := time.Now()
 	//	ctx, cancel := context.WithTimeout(c.ctx, pdTimeout)
@@ -355,6 +358,13 @@ func (c *client) processTSORequests(stream pdpb.PD_TsoClient, requests []*tsoReq
 		Header: c.requestHeader(),
 		Count:  uint32(len(requests)),
 	}
+
+	global++
+	if global > 1000 {
+		global = 0
+		fmt.Println("!!!!!!!!!!! ", req.Count)
+	}
+
 	if err := stream.Send(req); err != nil {
 		c.finishTSORequest(requests, 0, 0, err)
 		c.scheduleCheckLeader()
@@ -434,32 +444,32 @@ var tsoReqPool = sync.Pool{
 	},
 }
 
-// func (c *client) GetTS(ctx context.Context) (int64, int64, error) {
-// 	resp := c.GetTSAsync()
-// 	resp.wg.Wait()
-// 	return resp.physical, resp.logical, resp.err
-// }
-
 func (c *client) GetTS(ctx context.Context) (int64, int64, error) {
-	start := time.Now()
-	defer func() { cmdDuration.WithLabelValues("tso").Observe(time.Since(start).Seconds()) }()
-
-	req := tsoReqPool.Get().(*tsoRequest)
-	c.tsoRequests <- req
-
-	select {
-	case err := <-req.done:
-		if err != nil {
-			cmdFailedDuration.WithLabelValues("tso").Observe(time.Since(start).Seconds())
-			return 0, 0, errors.Trace(err)
-		}
-		physical, logical := req.physical, req.logical
-		tsoReqPool.Put(req)
-		return physical, logical, err
-	case <-ctx.Done():
-		return 0, 0, errors.Trace(ctx.Err())
-	}
+	resp := c.GetTSAsync()
+	resp.wg.Wait()
+	return resp.physical, resp.logical, resp.err
 }
+
+// func (c *client) GetTS(ctx context.Context) (int64, int64, error) {
+// 	start := time.Now()
+// 	defer func() { cmdDuration.WithLabelValues("tso").Observe(time.Since(start).Seconds()) }()
+
+// 	req := tsoReqPool.Get().(*tsoRequest)
+// 	c.tsoRequests <- req
+
+// 	select {
+// 	case err := <-req.done:
+// 		if err != nil {
+// 			cmdFailedDuration.WithLabelValues("tso").Observe(time.Since(start).Seconds())
+// 			return 0, 0, errors.Trace(err)
+// 		}
+// 		physical, logical := req.physical, req.logical
+// 		tsoReqPool.Put(req)
+// 		return physical, logical, err
+// 	case <-ctx.Done():
+// 		return 0, 0, errors.Trace(ctx.Err())
+// 	}
+// }
 
 func (c *client) GetRegion(ctx context.Context, key []byte) (*metapb.Region, *metapb.Peer, error) {
 	start := time.Now()
@@ -583,7 +593,7 @@ type asyncTSOClient struct {
 }
 
 func newAsyncTSOClient(cli pdpb.PD_TsoClient, header *pdpb.RequestHeader) *asyncTSOClient {
-	const chanSize = 2048
+	const chanSize = 1024
 	ret := &asyncTSOClient{
 		cli:    cli,
 		header: header,
@@ -609,18 +619,34 @@ func (async *asyncTSOClient) backgroundSendWorker() {
 	// Close workCh will notify the receive goroutine to exit.
 	defer close(async.workCh)
 
+	xxx := 0
+
 	for {
-		count := len(async.input)
-		batch := make([]*TSOResponse, count)
-		for i := 0; i < count; i++ {
-			tmp := <-async.input
-			batch[i] = tmp
+		first, ok := <-async.input
+		if !ok {
+			return
 		}
+
+		remain := len(async.input)
+		batch := make([]*TSOResponse, remain+1)
+		batch[0] = first
+		for i := 0; i < remain; i++ {
+			tmp := <-async.input
+			batch[i+1] = tmp
+		}
+		count := remain + 1
 
 		req := &pdpb.TsoRequest{
 			Header: async.header,
 			Count:  uint32(count),
 		}
+
+		xxx++
+		if xxx > 5000 {
+			xxx = 0
+			fmt.Println(" count ------------ ", count)
+		}
+
 		err := async.cli.Send(req)
 		if err != nil {
 			for i := 0; i < len(batch); i++ {