@tiancaiamao 2017-06-05
Wait a moment please, I want to test some variant of this code. @siddontang