@tiancaiamao 2017-06-06
After tune, this branch could reach 170w QPS, compare to 100w QPS of master branch,
and the latency is also lower than before.
I think it works!

PTAL @siddontang @disksing @coocood