@tiancaiamao 2017-06-06
After:    pd-tso-bench CPU 665%, pd-server CPU 86%
```
➜  [115] bin git:(25224fe) ✗ ./pd-tso-bench -pd "192.168.199.118:2379"       
INFO[0000] [pd] create pd client with endpoints [192.168.199.118:2379] 
INFO[0000] [pd] leader switches to: http://192.168.199.118:2379, previous:  
INFO[0000] [pd] init cluster id 6426875904493309296     
count:1766663, max:4, min:0, >1ms:59320, >2ms:5817, >5ms:0, >10ms:0, >30ms:0
count:1731660, max:3, min:0, >1ms:71218, >2ms:4917, >5ms:0, >10ms:0, >30ms:0
count:1769959, max:2, min:0, >1ms:63565, >2ms:6289, >5ms:0, >10ms:0, >30ms:0
count:1747821, max:3, min:0, >1ms:63644, >2ms:7303, >5ms:0, >10ms:0, >30ms:0
count:1735035, max:2, min:0, >1ms:72543, >2ms:4677, >5ms:0, >10ms:0, >30ms:0
count:1763472, max:2, min:0, >1ms:71101, >2ms:1830, >5ms:0, >10ms:0, >30ms:0
count:1794247, max:2, min:0, >1ms:57649, >2ms:2506, >5ms:0, >10ms:0, >30ms:0
count:1755183, max:2, min:0, >1ms:81964, >2ms:1438, >5ms:0, >10ms:0, >30ms:0
count:1737065, max:2, min:0, >1ms:73377, >2ms:6844, >5ms:0, >10ms:0, >30ms:0
count:1788889, max:2, min:0, >1ms:65929, >2ms:1331, >5ms:0, >10ms:0, >30ms:0
count:1722003, max:3, min:0, >1ms:75876, >2ms:6213, >5ms:0, >10ms:0, >30ms:
```

Before:  pd-tso-bench CPU 428%, pd-server CPU 30%
```
➜  [115] bin git:(25224fe) ✗ ./pd-tso-bench -pd "192.168.199.118:2379"
INFO[0000] [pd] create pd client with endpoints [192.168.199.118:2379] 
INFO[0000] [pd] leader switches to: http://192.168.199.118:2379, previous:  
INFO[0000] [pd] init cluster id 6426875904493309296     
count:1015093, max:6, min:0, >1ms:442941, >2ms:3007, >5ms:283, >10ms:0, >30ms:0
count:983313, max:3, min:0, >1ms:479283, >2ms:1999, >5ms:0, >10ms:0, >30ms:0
count:977092, max:2, min:0, >1ms:510326, >2ms:3, >5ms:0, >10ms:0, >30ms:0
count:1076943, max:3, min:0, >1ms:328533, >2ms:1869, >5ms:0, >10ms:0, >30ms:0
count:1045777, max:5, min:0, >1ms:377137, >2ms:627, >5ms:1, >10ms:0, >30ms:0
count:1009418, max:3, min:0, >1ms:458588, >2ms:1505, >5ms:0, >10ms:0, >30ms:0
count:1008807, max:2, min:0, >1ms:469728, >2ms:62, >5ms:0, >10ms:0, >30ms:0
count:1007306, max:3, min:0, >1ms:470365, >2ms:1645, >5ms:0, >10ms:0, >30ms:0
count:1016277, max:3, min:0, >1ms:451836, >2ms:350, >5ms:0, >10ms:0, >30ms:0
count:1003122, max:3, min:0, >1ms:474957, >2ms:1053, >5ms:0, >10ms:0, >30ms:0
count:990887, max:2, min:0, >1ms:494182, >2ms:467, >5ms:0, >10ms:0, >30ms:0
count:970245, max:2, min:0, >1ms:526198, >2ms:351, >5ms:0, >10ms:0, >30ms:0
count:985610, max:4, min:0, >1ms:499713, >2ms:2467, >5ms:0, >10ms:0, >30ms:0
count:1002671, max:2, min:0, >1ms:470846, >2ms:306, >5ms:0, >10ms:0, >30ms:0
count:1043497, max:3, min:0, >1ms:399441, >2ms:1459, >5ms:0, >10ms:0, >30ms:0
count:1022783, max:1, min:0, >1ms:434168, >2ms:0, >5ms:0, >10ms:0, >30ms:0
count:988805, max:5, min:0, >1ms:489818, >2ms:1632, >5ms:1, >10ms:0, >30ms:0
count:1003523, max:2, min:0, >1ms:471856, >2ms:221, >5ms:0, >10ms:0, >30ms:0
count:1004084, max:2, min:0, >1ms:475264, >2ms:751, >5ms:0, >10ms:0, >30ms:0
count:984307, max:4, min:0, >1ms:498610, >2ms:1034, >5ms:0, >10ms:0, >30ms:0
```

@coocood