@tiancaiamao 2017-06-07
Not yet.
This API is asynchronous so it won't make the caller block. @siddontang