@siddontang 2017-06-07
Other place also needs to call Wait, this can block.