@tiancaiamao 2017-06-08
Now pass a `ctx.WithTimeout()` would control the timeout.
PTAL @siddontang @disksing @coocood