@siddontang 2017-06-08
Any benchmark now? I worry too many timers will hurt the performance.