@tiancaiamao 2017-06-08
WithTimeout() slight affect the result, but I think it's still acceptable.
CPU goes from 665% to 695%.

```
➜  [115] bin git:(00efc37) ✗ ./pd-tso-bench -pd "192.168.199.118:2379"        
INFO[0000] [pd] create pd client with endpoints [192.168.199.118:2379] 
INFO[0000] [pd] leader switches to: http://192.168.199.118:2379, previous:  
INFO[0000] [pd] init cluster id 6426875904493309296     
count:1514209, max:9, min:0, >1ms:177816, >2ms:22633, >5ms:671, >10ms:0, >30ms:0
count:1556232, max:3, min:0, >1ms:157221, >2ms:20917, >5ms:0, >10ms:0, >30ms:0
count:1610625, max:2, min:0, >1ms:158271, >2ms:10017, >5ms:0, >10ms:0, >30ms:0
count:1585037, max:4, min:0, >1ms:155283, >2ms:21495, >5ms:0, >10ms:0, >30ms:0
count:1589266, max:3, min:0, >1ms:164893, >2ms:10394, >5ms:0, >10ms:0, >30ms:0
count:1537842, max:4, min:0, >1ms:171930, >2ms:22383, >5ms:0, >10ms:0, >30ms:0
count:1550060, max:3, min:0, >1ms:173469, >2ms:20129, >5ms:0, >10ms:0, >30ms:0
count:1554103, max:4, min:0, >1ms:159501, >2ms:18023, >5ms:0, >10ms:0, >30ms:0
count:1544470, max:4, min:0, >1ms:165368, >2ms:18989, >5ms:0, >10ms:0, >30ms:0
count:1562577, max:3, min:0, >1ms:172731, >2ms:16997, >5ms:0, >10ms:0, >30ms:0
count:1554627, max:3, min:0, >1ms:165342, >2ms:16671, >5ms:0, >10ms:0, >30ms:0
count:1529211, max:7, min:0, >1ms:164869, >2ms:20365, >5ms:999, >10ms:0, >30ms:0
count:1533072, max:3, min:0, >1ms:180492, >2ms:13687, >5ms:0, >10ms:0, >30ms:0
```

@siddontang