@siddontang 2017-06-08
But QPS reduces from 1722003 to 1533072, about 10%.