@tiancaiamao 2017-06-08
Through timeout context slightly hurt the performance, it still beat the request-response way.

1. In the benchmark battle, new branch shows better QPS, and with a larger percentage latency < 1ms.
2. In the real world case, there won't have that much pressure, so the request-response way batch count reduce and perform poor, while streaming way won't be affected.

If I test with 500 connection, and sleep 5ms, the result is as follows:

Before:
```
➜  [115] bin git:(00efc37) ✗ ./pd-tso-bench -pd "192.168.199.118:2379" -C 500 -sleep 5ms
INFO[0000] [pd] create pd client with endpoints [192.168.199.118:2379] 
INFO[0000] [pd] leader switches to: http://192.168.199.118:2379, previous:  
INFO[0000] [pd] init cluster id 6426875904493309296     
count:574993, max:2, min:0, >1ms:106863, >2ms:427, >5ms:0, >10ms:0, >30ms:0
count:609093, max:2, min:0, >1ms:60458, >2ms:157, >5ms:0, >10ms:0, >30ms:0
count:628347, max:3, min:0, >1ms:89517, >2ms:4, >5ms:0, >10ms:0, >30ms:0
count:746333, max:2, min:0, >1ms:1377, >2ms:171, >5ms:0, >10ms:0, >30ms:0
count:665922, max:2, min:0, >1ms:25801, >2ms:97, >5ms:0, >10ms:0, >30ms:0
count:635950, max:1, min:0, >1ms:22853, >2ms:0, >5ms:0, >10ms:0, >30ms:0
count:631592, max:3, min:0, >1ms:31806, >2ms:207, >5ms:0, >10ms:0, >30ms:0
count:607635, max:2, min:0, >1ms:59624, >2ms:492, >5ms:0, >10ms:0, >30ms:0
count:599039, max:1, min:0, >1ms:66919, >2ms:0, >5ms:0, >10ms:0, >30ms:0
count:627326, max:2, min:0, >1ms:31910, >2ms:259, >5ms:0, >10ms:0, >30ms:0
count:622266, max:2, min:0, >1ms:32692, >2ms:103, >5ms:0, >10ms:0, >30ms:0
count:591833, max:1, min:0, >1ms:70430, >2ms:0, >5ms:0, >10ms:0, >30ms:0
count:598753, max:3, min:0, >1ms:63633, >2ms:502, >5ms:0, >10ms:0, >30ms:0
count:592854, max:2, min:0, >1ms:76142, >2ms:39, >5ms:0, >10ms:0, >30ms:0
count:591635, max:3, min:0, >1ms:71246, >2ms:396, >5ms:0, >10ms:0, >30ms:0
count:590579, max:2, min:0, >1ms:63295, >2ms:619, >5ms:0, >10ms:0, >30ms:0
```

After:
```
➜  [115] bin git:(00efc37) ✗ ./pd-tso-bench -pd "192.168.199.118:2379" -C 500 -sleep 5ms 
INFO[0000] [pd] create pd client with endpoints [192.168.199.118:2379] 
INFO[0000] [pd] leader switches to: http://192.168.199.118:2379, previous:  
INFO[0000] [pd] init cluster id 6426875904493309296     
count:1251201, max:4, min:0, >1ms:74651, >2ms:1554, >5ms:0, >10ms:0, >30ms:0
count:1156134, max:2, min:0, >1ms:83213, >2ms:1902, >5ms:0, >10ms:0, >30ms:0
count:1258752, max:4, min:0, >1ms:71080, >2ms:549, >5ms:0, >10ms:0, >30ms:0
count:1201618, max:2, min:0, >1ms:85059, >2ms:1287, >5ms:0, >10ms:0, >30ms:0
count:1243404, max:5, min:0, >1ms:79653, >2ms:1940, >5ms:83, >10ms:0, >30ms:0
count:1261740, max:2, min:0, >1ms:73828, >2ms:469, >5ms:0, >10ms:0, >30ms:0
count:1267599, max:2, min:0, >1ms:80703, >2ms:376, >5ms:0, >10ms:0, >30ms:0
count:1228513, max:2, min:0, >1ms:89393, >2ms:1252, >5ms:0, >10ms:0, >30ms:0
count:1239034, max:2, min:0, >1ms:82583, >2ms:2877, >5ms:0, >10ms:0, >30ms:0
count:1193033, max:2, min:0, >1ms:93077, >2ms:2589, >5ms:0, >10ms:0, >30ms:0
count:1242565, max:3, min:0, >1ms:84474, >2ms:1799, >5ms:0, >10ms:0, >30ms:0
count:1208362, max:2, min:0, >1ms:84238, >2ms:1262, >5ms:0, >10ms:0, >30ms:0
count:1156218, max:2, min:0, >1ms:93749, >2ms:1702, >5ms:0, >10ms:0, >30ms:0
count:1169839, max:4, min:0, >1ms:90786, >2ms:5344, >5ms:0, >10ms:0, >30ms:0
count:1198327, max:2, min:0, >1ms:91646, >2ms:4367, >5ms:0, >10ms:0, >30ms:0
count:1224089, max:2, min:0, >1ms:90332, >2ms:1823, >5ms:0, >10ms:0, >30ms:0
count:1233253, max:2, min:0, >1ms:94360, >2ms:717, >5ms:0, >10ms:0, >30ms:0
```